Práctica obligatoria

+++ install_db.js
@@ -1,4 +1,6 @@
+/* eslint-disable no-console */
 const mongoose = require('mongoose')
+const readline = require('readline')
 
 const { Ad, User } = require('./models')
 
@@ -32,12 +34,35 @@ const loadDefaultData = async () => {
 }
 
 const main = async () => {
-  // TODO: Asegurar BBDD
+  
   await mongoose.connect(MONGODB_URI)
   console.info(`Connected to MongoDB, at ${mongoose.connection.name} database`)
+
+  if (!(await askYesNo('Estas seguro que quieres inicializar la BD? (Si/No)\n'))){
+    console.log('Se ha cancelado la inicialización de la base de datos.')
+    return process.exit(0)
+  }
+
   await deleteAllData()
   await loadDefaultData()
   mongoose.connection.close()
+}
+
+function askYesNo(questionText) {
+  return new Promise((resolve) => {
+    const rl = readline.createInterface({
+      input: process.stdin,
+      output: process.stdout,
+    })
+    rl.question(questionText, (answer) => {
+      rl.close()
+      if (answer.toLowerCase() === 'si') {
+        resolve(true)
+        return
+      }
+      resolve(false)
+    })
+  })
 }
 
 main().catch((err) => console.log(err))
